Move the threats state to a new component
